Fix emacs native compilation warning
```
 ■  Warning (comp): pfuture.el:61:2: Warning: docstring has wrong usage
of unescaped single quotes (use \= or different quoting)
 ■  Warning (comp): pfuture.el:123:2: Warning: docstring has wrong usage
of unescaped single quotes (use \= or different quoting)
```

To reproduce, have emacs built with native compilation and notice the
compilation logs. You can then open a file and run `M-x
emacs-lisp-native-compile-and-load` before and after the changes to see the
warning is removed.